Add REST API for SuperTuxKart

+++ src/config/player_profile.cpp
@@ -157,7 +157,7 @@ void PlayerProfile::addIcon()
     if (m_icon_filename.size() > 0 || isGuestAccount())
         return;
 
-    int n = (m_unique_id + kart_properties_manager->getKartId("tux") - 1)
+    size_t n = (m_unique_id + kart_properties_manager->getKartId("tux") - 1)
           % kart_properties_manager->getNumberOfKarts();
 
     std::string source = kart_properties_manager->getKartById(n)
@@ -285,4 +285,3 @@ bool PlayerProfile::operator<(const PlayerProfile &other)
 }   // operator<
 
 // -----------------------------------------------------------------------------
-
